+++ src/popup/main.js
@@ -1,21 +1,19 @@
 require("../styles/popup_entry.less")
 var clock = require("./clock")
 
-/**
- *  window event
- */
-window.onload = () => {
+
+document.querySelector('#refresh_button').addEventListener('click', clock.reset)
+
+document.querySelector('#options_button').addEventListener('click', e => {
+    e.preventDefault()
+    browser.runtime.openOptionsPage()
+})
+
+window.addEventListener("load", function (event) {
     browser.runtime.onMessage.addListener(clock.update)
     clock.request()
+})
 
-    document.querySelector('#refresh_button').addEventListener('click', clock.reset)
-
-    document.querySelector('#options_button').addEventListener('click', e => {
-        e.preventDefault()
-        browser.runtime.openOptionsPage()
-    })
-}
-
-window.onunload = () => {
+window.addEventListener("unload", function (event) {
     browser.runtime.onMessage.removeListener(clock.update)
-}
+})